requirements.txt
sentence-transformers
numpy
scikit-learn
google-generativeai
torch
onnxruntime
optimum
onnx
